+++ src/components/Buttons/PlaytestButton.js
@@ -1,4 +1,4 @@
-import React from "react"
+import React, {useState} from "react"
 import ClickToOpen from "./ClickToOpen"
 import buttonClasses from "./ClickToOpen.module.css"
 import * as Constants from "../../constants";
@@ -7,7 +7,9 @@ export default function PlaytestButton() {
 
     const apiUrlFunction = (email) => Constants.PLAYTEST_API + "&Message=" + email;
 
-    const playtestButton = <ClickToOpen variant="warning" size={buttonClasses.independentBlack} 
+    const [show, setShow] = useState(false);
+
+    const playtestButton = <ClickToOpen variant="warning" size={buttonClasses.independentBlack} show={show} setShow={setShow}
                                 text={Constants.PLAYTEST_BUTTON_TEXT} title={Constants.PLAYTEST_MODAL_TITLE}
                                 apiUrlFunction={apiUrlFunction} apiKey={Constants.PLAYTEST_API_KEY}/>
