+++ src/components/MyForm.js
@@ -1,11 +1,12 @@
 import React from 'react';
-import {ErrorMessage, Formik, Field, Form } from 'formik';
+import {Formik, Form } from 'formik';
 import {Button, Row, Col} from 'react-bootstrap';
-import styled from "styled-components";
 import * as Yup from 'yup';
 import Break from "../components/Break";
 import axios from "axios";
- 
+import Email from "./FormFields/Email";
+import CheckHuman from "./FormFields/CheckHuman";
+import FormFooter from "./FormFields/FormFooter";
 
 
 export default function MyForm(props) {
@@ -14,26 +15,7 @@ export default function MyForm(props) {
     email: Yup.string().email('Invalid email').required('Required'),
     test: Yup.number('Invalid').min(4, 'Incorrect').max(4, 'Incorrect').required('You must pass the test'),
   });
-
-
-  const MyLabel = styled.p`
-    font-size: 20px;
-    display: block;
-  `
-
-  const MyField = styled(Field)`
-    margin-bottom: 25px;
-  `
-
-  const FullField = styled(MyField)`
-    width: 100%;
-  `
-
-  const TestQuestion = styled.p`
-    margin-left: 10px;
-  `
-
-
+  
   return  <div>
             <Formik
               initialValues={{
@@ -52,37 +34,11 @@ export default function MyForm(props) {
             }
             >
             <Form>
-                <MyLabel>Email</MyLabel>
-                <Row>
-                  <Col>
-                    <FullField id="email" name="email" placeholder="Your email address" />
-                    <ErrorMessage name="email" />
-                  </Col>
-                </Row>
-                <Break height="3"/>
-                <MyLabel>Are you undead?</MyLabel>
-                <Row>
-                  <TestQuestion>2 + 2 =</TestQuestion>
-                  <Col>
-                    <FullField
-                      id="test"
-                      name="test"
-                      placeholder="What is the answer?"
-                      type="text"
-                    />
-                    <ErrorMessage name="test"/>
-                  </Col>
-                </Row>
-              <Row>
-                <Col xs="1">
-                    <Button variant="secondary" onClick={props.close}>Close</Button>
-                </Col>
-                <Col xs="7" sm="8" md="8" lg="9" xlg="10"></Col>
-                <Col xs="1">
-                    <Button variant="primary" type="submit">Onwards</Button>
-                </Col>
-              </Row>
-              </Form>
+              <Email/>
+              <Break height="3"/>
+              <CheckHuman/>
+              <FormFooter close={props.close}/>
+             </Form>
             </Formik>
           </div>
 };
